Create Monte Carlo Simulation
To create this simulation I started out by creating a dictionary with 5 keys and set all their values to zero and an accumulation variable for the days set to 0. The while loop I wrote iterates until one of the dictionaries values becomes two (representing earning a second coupon) and after each iteration of the loop the days value is incremented by 1. I put the range in 10,000 to simulate this experiment 10,000 times and every result of days was appended to an empty list called res. I then took the average of my res list to find the expected value of 10,000 simulations of this experiments for maximum assurance.