Create Matrices and Vectors
Here are created a function that takes 3 arrays as input values and then stacks them into matrix using the numpy method vstack. I then use np and matrix methods to take the rank of the matrix, the rank of the matrix's transposition and the determinant of the matrix, respectively. I then created a nested function called matrix_first_el that consists of an if statement that states if the matrix determinant is not zero, then return the first element (first value in first row) of the matrix rounded to the nearest hundredth. If the matrix determinant does equal zero then return 'no inverse' because if the determinant of a matrix is zero then it's inverse does not exist.